refactor(reader): move into modules
